refactor: insere correção de nomenclatura pois models estão relacionados a banco
