+++ Arbooks.Web/Views/Home/BookDetails.cshtml
@@ -1,5 +1,5 @@
-﻿@using Arbooks.Web.Models
-@model Book
+﻿@using Arbooks.Web.DTOs
+@model BookDTO
 
 @{
     ViewData["Title"] = "Details";
