+++ Arbooks.Web/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿﻿@using Arbooks.Web
-@using Arbooks.Web.Models
+@using Arbooks.Web.DTOs
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
